refactor: Disease 서비스 구현체에 대한 인터페이스 생성
